+++ resources/release-notes.md
@@ -1,4 +1,5 @@
 # Software
 
 - Add support for compression flag appearing in CSV files
+- Add support for temperature calibration of FEE1 and FEE2
 - Add compression flag to field plot, if no other event is requested and compression flag changes